Move the ValueBuilder stuff from value.h to value_builder_factory.h as that is required in far fewer places than value.h.
