+++ eval/src/vespa/eval/eval/fast_value.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include "value.h"
+#include "value_builder_factory.h"
 
 namespace vespalib::eval {
