+++ eval/src/vespa/eval/eval/value_codec.h
@@ -13,6 +13,8 @@ namespace vespalib::eval {
 
 VESPA_DEFINE_EXCEPTION(DecodeValueException, Exception);
 
+class ValueBuilderFactory;
+
 /**
  * encode a value (which must support the new APIs) to binary format
  **/
