+++ eval/src/vespa/eval/instruction/generic_merge.cpp
@@ -2,11 +2,11 @@
 
 #include "generic_merge.h"
 #include <vespa/eval/eval/inline_operation.h>
+#include <vespa/eval/eval/value_builder_factory.h>
 #include <vespa/eval/eval/wrap_param.h>
 #include <vespa/vespalib/util/stash.h>
 #include <vespa/vespalib/util/typify.h>
 #include <cassert>
-#include <typeindex>
 
 using namespace vespalib::eval::tensor_function;
